Textures
- fixed complex texture mapping

+++ shaders/lab2/fsh.glsl
@@ -3,7 +3,7 @@
 uniform sampler2DArray u_surfaceMap; // Текстуры поверхности
 
 in vec2 v_texCoord;
-flat in int v_texNum;
+flat in float v_texNum;
 
 layout(location = 0) out vec4 o_color;
